- New Aircraft struct with per-phase speeds, fuel data, FuelType enum
- 14 TOML aircraft profiles in data/aircraft/ (GA through wide-body)
- Build.rs pipeline: parse TOML, validate, generate aircraft_db.rs
- Runtime custom profile loading via --profile flag
- LNM .lnmperf import subcommand (XML + INI formats)
- FORMAT.md schema documentation for human and AI authoring
- CLI restructured: aircraft list, aircraft import --format lnmperf

+++ src/error.rs
@@ -14,8 +14,8 @@ pub enum Error {
     #[error("unknown airport ICAO code: {icao}")]
     UnknownAirport { icao: String },
 
-    #[error("flight distance {distance_nm:.0} nm exceeds aircraft range of {range_nm} nm")]
-    RangeExceeded { distance_nm: f64, range_nm: u32 },
+    #[error("flight distance {distance_nm:.0} nm exceeds aircraft range of {range_nm:.0} nm")]
+    RangeExceeded { distance_nm: f64, range_nm: f64 },
 
     #[error("runway at {airport_icao} is {available_ft} ft, aircraft requires {required_ft} ft")]
     RunwayTooShort {
